bzlib.command - nix Close, Fix, Reopen, Resolve; add Status
Remove the Close, Fix, Reopen and Resolve commands, and replace
with a single Status command.  Shortcuts for the former commands
should be implemented using aliases.

Motivated by the fact that the Bugzilla workflow - both statuses
and resolutions - is configurable.  Perhaps it is worthwhile to
provide the a default config with aliases for the default statuses
and resolutions.

Eventually, per-server aliases should be implemented, since
bugzilla is designed to work with multiple Bugzilla servers.