bzlib - store version info in bzlib.version
So that version information is available in bzlib, and so that there
is (still) exactly one place where version information is kept,
add a 'version' attribute to the bzlib module.

setup.py is able to access this property since sys.path[0] is the
directory of the invoked script - the correct directory in this case.

Also add a global --version argument to the bugzilla program.

Also bump version to 0.2dev.