bzr - initial commit of bugzillatools Bazaar plugin
This is a copy & paste of files from a Bazaar repository.  Couldn't be
bothered with a Git-Bazaar gateway, so I'm throwing out the history.

Let it be known that I don't care much for Bazaar (so slooooooow!), but
I use it at work and I wrote the Bugzilla plugin for it, so it made sense
to include in bugzillatools.

Porting to bzlib is yet to be done (it previously used a very basic web
mechaniser; that module was not imported).