bzr - fix circular import and remove fixes module
hooks and commands were recursively importing each other; since there
is just one small command, define it in the hook method itself and remove
the commands module.

Extend the type() of the commit command if the command is an instance of
bzrlib.builtins.cmd_commit, rather than extending same, in case the
command is already a subclass (i.e. from some other plugin).

Because all hooks and commands are now defined in the same file, curry
the post-commit hook with the bugs and remove the fixes module.