bzlib.bugzilla - restrict field visibility according to bug data
The allowable values of some fields are restricted to some subset of all
the possible values according to the current value of some other field.

Add to ``bzlib.bugzilla.Bugzilla.get_field_values`` an optional
``visible_for`` argument, which should be a dict of the current bug data.
If supplied, and if the requested field restricts its values based on
the value of a field that is present in the given dict, return only those
fields that are visible for the current value of that field.