bzlib.command - new command New: file new bugs
The ``new`` command creates a new bug, prompting user for values to
fields.  At this stage, only mandatory fields are prompted for.

The command is slightly intelligent as to whether to prompt user to
choose from a set of values or ask for arbitrary text input, but it
will get better in future.