editor: fix encoding errors
If the argument to editor.input is a unicode string, other strings
are coerced to unicode and the string is written to the temporary
file using the internal representation (e.g. UCS-4/UTF-32),
resulting in ugly comments and possibly invalid XML sent to Bugzilla
server.

Avoid this situation by encoding the argument as a utf-8 bytestring.