Load/save cookies in a supported way.
The SimpleCookie class is not supposed to accept a header ("Set-Cookie:") when
loading from a string.

See https://bugs.python.org/issue22758#msg230205 for more information

We were relying on that non-standard way of cookie loading to save and restore
our cookies. This bug appears to have been fixed in python3.5.

This commit will allow Bugzillatools to continue saving/loading cookies
properly on python3.5 and onwards.