Build transport before creating XML ServerProxy
Python 3.3's xmlrpc.client.ServerProxy doesn't recognize `context` as
one of the constructor parameters. In order to support the use of custom
SSL contexts when connecting to Bugzilla, we must build the transport
and set the SSL context on it first before constructing our ServerProxy
instead of passing the context directly to the ServerProxy constructor.